+++ packages/vchart/src/component/axis/cartesian/axis.ts
@@ -412,7 +412,7 @@ export abstract class CartesianAxis<T extends ICartesianAxisCommonSpec = ICartes
     return scales;
   }
 
-  protected collectData(depth?: number) {
+  protected collectData(depth?: number, rawData?: boolean) {
     const data: { min: number; max: number; values: any[] }[] = [];
     eachSeries(
       this._regions,
@@ -438,8 +438,12 @@ export abstract class CartesianAxis<T extends ICartesianAxisCommonSpec = ICartes
         const seriesData = s.getViewDataStatistics?.();
         if (field) {
           field.forEach(f => {
-            if (seriesData?.latestData?.[f]) {
-              data.push(seriesData.latestData[f]);
+            if (rawData) {
+              data.push(s.getRawDataStatisticsByField(f, false) as { min: number; max: number; values: any[] });
+            } else {
+              if (seriesData?.latestData?.[f]) {
+                data.push(seriesData.latestData[f]);
+              }
             }
           });
         }
@@ -474,6 +478,26 @@ export abstract class CartesianAxis<T extends ICartesianAxisCommonSpec = ICartes
       }
     );
   }
+
+  // protected _seriesUpdateAfterScaleChange(updateInfo: { domain?: boolean; range?: boolean; type?: boolean }) {
+  //   const orient = this.getOrient();
+  //   eachSeries(
+  //     this._regions,
+  //     s => {
+  //       if (isXAxis(orient)) {
+  //         (s as ICartesianSeries).xAxisUpdated(updateInfo);
+  //       } else if (isYAxis(orient)) {
+  //         (s as ICartesianSeries).yAxisUpdated(updateInfo);
+  //       } else if (isZAxis(orient)) {
+  //         (s as ICartesianSeries).zAxisUpdated(updateInfo);
+  //       }
+  //     },
+  //     {
+  //       userId: this._seriesUserId,
+  //       specIndex: this._seriesIndex
+  //     }
+  //   );
+  // }
 
   _transformLayoutPosition = (pos: Partial<IPoint>) => {
     let { x, y } = pos;